+++ R/helpers-testing.R
@@ -61,7 +61,7 @@ retrieve_credential_testing <- function(
   }
 
   retrieve_credential_local(
-    path_credential = path_credential, # "misc/dev-2.credentials"
+    path_credential = path_credential, # "misc/example.credentials"
     project_id      = d_map$project_id,
     username        = username
   )