+++ R/retrieve-credential.R
@@ -117,7 +117,7 @@
 #' @examples
 #' \dontrun{
 #' # ---- Local File Example ----------------------------
-#' path <- system.file("misc/dev-2.credentials", package = "REDCapR")
+#' path <- system.file("misc/example.credentials", package = "REDCapR")
 #' (p1  <- REDCapR::retrieve_credential_local(path, 33L))
 #' (p2  <- REDCapR::retrieve_credential_local(path, 34L))
 #'